[add] fields to car class and restructured
