--- src/main/java/model/Car.java
@@ -1,21 +1,0 @@
-package model;
-
-import lombok.*;
-
-import javax.persistence.*;
-import java.util.List;
-import java.util.Set;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-public class Car {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private CarType type;
-}
